Application_server_Dockerized/Model_Server/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
matplotlib
numpy
python-multipart